pdf-organize/requirements.txt
google-generativeai
PyMuPDF
watchdog
python-dotenv
